LayoutTests/fast/mediastream/RTCPeerConnection-createOffer-expected.txt

Tests RTCPeerConnection createOffer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.createOffer(requestSucceeded1, requestFailed1); did not throw exception.
PASS createOffer request succeeded.
PASS sessionDescription.type is "offer"
FAIL pc.createOffer(requestSucceeded2, requestFailed2, {}); should throw an exception. Was [object Promise].
FAIL createOffer request succeeded.
PASS successfullyParsed is true

TEST COMPLETE
